functions/deprovisioning-function/pyproject.toml
[project]
name = "deprovisioning-function"
version = "0.1.0"
description = "Azure Function for automated deprovisioning of feature branch environments"
requires-python = ">=3.11"
dependencies = [
    "azure-functions",
    "azure-identity",
    "azure-mgmt-resource",
    "azure-mgmt-containerservice",
    "azure-mgmt-rdbms",
    "azure-mgmt-msi",
    "kubernetes",
    "pydantic",
    "pydantic-settings",
    "psycopg2-binary",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
    "mypy",
    "types-requests",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "C", "I", "N", "D", "UP", "ANN", "S", "B", "A", "COM", "DTZ", "EM", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SIM", "TCH", "ARG", "PTH", "ERA", "PD", "PGH", "PL", "TRY", "NPY", "RUF"]
ignore = ["D100", "D101", "D102", "D103", "D104", "D105", "ANN101", "ANN102"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
